core.fsync: add `fsync_component()` wrapper which doesn't die
We have a `fsync_component_or_die()` helper function which only syncs
changes to disk in case the corresponding config is enabled by the user.
This wrapper will always die on an error though, which makes it
insufficient for new callsites we are about to add.

Add a new `fsync_component()` wrapper which returns an error code
instead of dying.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>